Move compare count transformer
